Work in progress. Separate code into LoginForm and RegistrationForm components, and AuthPage page
